requirements.txt
Django==3.2.13
docker_py==1.10.6
kubernetes==25.3.0
MySQL-python==1.2.5
pandas==1.2.3
Pillow==9.4.0
pycryptodome==3.16.0
pytz==2021.1
redis==4.4.0
requests==2.28.1
rsa==4.9
